Allow duplicate resources in the same biome group
Allows duplicate resources in the resourcequeue even if they are marked
as unique in the versionconfig.xml. It is only possible to add multiple
unique resources in a list in the same biome group, merging behaviours
with other biome groups and default values have not changed.

This means that you could import a world, create a biomegroup with all
biomes and configure multiple Ore(GOLD_ORE resources, these would then
replace all existing Ore(GOLD_ORE resources, but not any other resources
(if the merge option is chosen).