1.0.10: OTG pt1 (wip). Schematic > BO3 CustomObject
- Added all the required files for running TCEE in the build folder
- Added option for converting schematics to CustomObject BO3 (all blocks
in one BO3 instead of 16x16 BO3's with branches).
- For CustomObject BO3's added the option to set a block as the center
of a schematic.
- Added the option to remove air from BO3's.
- When converting schematics to BO3's, ntb files now get BO3 name + xyz
coordinate as a part of their filename, making it easier to find the nbt
file you need.
- Updated the comments in BO3's.
- Updated the UI a bit for OTG, cleaned up some clutter, removed the old
logo.
- Improved nearly all the message boxes / popups.
- Removed pop up boxes with warnings when editing the resourcequeue and
creating duplicate resources. Merging rules are only used when
generating configs and biome groups are merged, not while editing a
biome group in the biomes tab.
- Fixed some minor bugs like invalid names being allowed for biome
groups etc.

Still need to update the versionconfig.xml's for OTG 1.10.2 and 1.11.2,
all the new settings are still missing.